interpreter: modified calling convention to use stubs
In dynamic languages, call instructions need to adapt between caller and
callee by adding or removing arguments. This includes dropping the receiver
for a call to a field closure, adding bound arguments, dropping excess
arguments, and adding nil padding.

The call instruction must not modify arguments in the caller's frame, since
this would invalidate the caller's safe points. Instead, if any adjustment
is needed, the call instruction will create an extra stack frame and put
the adjusted arguments there. The GC marker can skip that frame, since it
will only contain values copied from the caller.

When an extra frame is used like this, the callee returns by invoking a
stub that pops both frames and moves the return values to the correct
place. The stub to invoke is encoded into the return address.

Related changes:

- The callnamedpropwithprototype instruction is removed. callnamedprop now
  expects the slot before the receiver to contain the value from which the
  property should be loaded (either the receiver or superclass).
- lox: fixed a bug where "this" couldn't be captured.
- lua: callee now gets passed to metamethod as first argument.